refactor: centralize admin company filter in sidebar

+++ app/Helper/helper.php
@@ -6,7 +6,7 @@ if(!function_exists('getCompanyId')){
     function getCompanyId($user): ?int
     {
         if ($user->role_id == Role::ADMIN) {
-            return session('company_id', 1);
+            return (int) session('company_id', 1);
         } elseif ($user->role_id == Role::COMPANY && $user->company) {
             return $user->company->id;
         } elseif ($user->role_id == Role::BRANCH && $user->branch) {
